requirements.txt
# Dependencias para el Analizador de Puntos de Minería
flask==3.1.2
pillow>=10.0.0
easyocr>=1.7.0
gunicorn==21.2.0
